deploy using settings directive is implemented
