+++ src/app/frontend/deploy/deploy_controller.js
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 import {stateName as zerostate} from 'zerostate/zerostate_state';
-import {stateName as replicasetliststate} from 'replicasetlist/replicasetlist_state';
 import showNamespaceDialog from 'deploy/createnamespace_dialog';
 
 /**
@@ -34,25 +33,6 @@ export default class DeployController {
     /** @export {string} */
     this.name = '';
 
-    /** @export {string} */
-    this.containerImage = '';
-
-    /** @export {number} */
-    this.replicas = 1;
-
-    /** @export {string} */
-    this.description = '';
-
-    /**
-     * List of supported protocols.
-     * TODO(bryk): Do not hardcode the here, move to backend.
-     * @const @export {!Array<string>}
-     */
-    this.protocols = ['TCP', 'UDP'];
-
-    /** @export {!Array<!backendApi.PortMapping>} */
-    this.portMappings = [this.newEmptyPortMapping_(this.protocols[0])];
-
     /**
      * List of available namespaces.
      * @export {!Array<string>}
@@ -64,9 +44,6 @@ export default class DeployController {
      * @export {(string|undefined)}
      */
     this.namespace = this.namespaces.length > 0 ? this.namespaces[0] : undefined;
-
-    /** @export {boolean} */
-    this.isExternal = false;
 
     /** @private {!angular.$resource} */
     this.resource_ = $resource;
@@ -82,42 +59,23 @@ export default class DeployController {
 
     /** @private {boolean} */
     this.isDeployInProgress_ = false;
+
+    /**
+     * Contains the selected directive's controller which has its own deploy logic
+     *
+     * @export {{deploy:function()}}
+     */
+    this.detail = {deploy: function() {}};
   }
 
   /**
-   * Deploys the application based on the sate of the controller.
+   * Notifies the child scopes to call their deploy methods.
    *
    * @export
    */
-  deploy() {
-    // TODO(bryk): Validate input data before sending to the server.
-
-    /** @type {!backendApi.AppDeploymentSpec} */
-    let appDeploymentSpec = {
-      containerImage: this.containerImage,
-      isExternal: this.isExternal,
-      name: this.name,
-      description: this.description,
-      portMappings: this.portMappings.filter(this.isPortMappingEmpty_),
-      replicas: this.replicas,
-      namespace: this.namespace,
-    };
-
-    /** @type {!angular.Resource<!backendApi.AppDeploymentSpec>} */
-    let resource = this.resource_('/api/appdeployments');
-
+  deployBySelection() {
     this.isDeployInProgress_ = true;
-    resource.save(
-        appDeploymentSpec,
-        (savedConfig) => {
-          this.isDeployInProgress_ = false;
-          this.log_.info('Successfully deployed application: ', savedConfig);
-          this.state_.go(replicasetliststate);
-        },
-        (err) => {
-          this.isDeployInProgress_ = false;
-          this.log_.error('Error deploying application:', err);
-        });
+    this.detail.deploy().finally(() => { this.isDeployInProgress_ = false; });
   }
 
   /**
@@ -157,25 +115,4 @@ export default class DeployController {
    * @export
    */
   cancel() { this.state_.go(zerostate); }
-
-  /**
-   * @param {string} defaultProtocol
-   * @return {!backendApi.PortMapping}
-   * @private
-   */
-  newEmptyPortMapping_(defaultProtocol) {
-    return {
-      port: null,
-      targetPort: null,
-      protocol: defaultProtocol,
-    };
-  }
-
-  /**
-   * Returns true when the given port mapping hasn't been filled by the user, i.e., is empty.
-   * @param {!backendApi.PortMapping} portMapping
-   * @return {boolean}
-   * @private
-   */
-  isPortMappingEmpty_(portMapping) { return !!portMapping.port && !!portMapping.targetPort; }
 }
